backend/requirements.txt
fastapi
uvicorn[standard]
openai
pypdf
youtube-transcript-api
python-dotenv
google-api-python-client
python-multipart
numpy
